+++ rombo/optimization/stdbo.py
@@ -8,7 +8,7 @@ from ..interpolation.interpolation import BoTorchModel
 
 # Setting data type and device for Pytorch based libraries
 tkwargs = {
-    "dtype": torch.float,
+    "dtype": torch.float64,
     "device": torch.device("cuda:0" if torch.cuda.is_available() else "cpu"),
 }
